requirements.txt
fastapi
bcrypt
pyjwt
python-dotenv
uvicorn
cryptography
sqlalchemy
psycopg2-binary
